+++ modules/data_analysis/builtin/include/cumsumBuiltin.hpp
@@ -15,7 +15,7 @@
 namespace Nelson::DataAnalysisGateway {
 //=============================================================================
 ArrayOfVector
-minBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+cumsumBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================