+++ modules/data_analysis/src/cpp/Minimum.cpp
@@ -456,7 +456,7 @@ Minimum(bool omitNaN, const ArrayOf& A, const ArrayOf& B, bool& needToOverload)
     }
     if (A.isSparse() || B.isSparse()) {
         needToOverload = true;
-        return ArrayOf();
+        return {};
     }
     Dimensions ASize(A.getDimensions());
     Dimensions BSize(B.getDimensions());