+++ CHANGELOG.md
@@ -40,6 +40,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
   - `isletter`: Determine which characters are letters.
 
+- data analysis module:
+
+  - `cumsum`: Cumulative sum.
+  - `cumprod`: Cumulative product.
+
 ### Changed
 
 - [#715](http://github.com/Nelson-numerical-software/nelson/issues/715): update `{fmt}` to 9.1.0.